Wifi: Add support for repeater mode
Repeater Mode is a feature where STA
network is shared to peers devices via
hotspot.
As part of this implementation introduce
a variable in WifiConfiguration to check
whether connected AP will share the
credentials with hotspot or not. If this
value is set to true then hotspot will turn
on using the ssid and passphrase of the
connected AP.
Expose one public api to check whether extended
wifi coverage setting is set or not.

CRs-Fixed: 2176801
Change-Id: I1f1f414f172173a6ccc4895769799742fa8fce02